create useBookMark hook to handle book marking a video
